Allow for breaking change in tokio-postgres v0.7.11
It seems that sqllogictest-rs defines this dependency:

```
tokio-postgres = { version = "0.7" }
```

which works up until 0.7.10. But at 0.7.11 tokio-postgres added an element to
[enum SimpleQueryMessage](https://docs.rs/tokio-postgres/0.7.11/src/tokio_postgres/lib.rs.html#246-255),
which is a breaking change. And since sqllogictest didn’t pin a bug revision
(which is ok, as long as it’s not major version v0), v0.7.11 actually represents
a breaking change rather than a bug fix.